X00TD: Add AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Signed-off-by: Vivekachooz <vivekachooz@gmail.com>
Change-Id: I40026d8685fccbcd4b7ccfa9f052e836ce483fd3